requirements.txt
fastapi
pydantic
twilio
gradio_client
uvicorn
python-decouple
python-multipart
cloudinary